Update vaccination scheduling to accurately reflect free vaccine schedules
Refactor VaccinationScheduleTemplateDAO to use double for age in months, enabling more precise age-based filtering and appointment generation. Update VaccinationService to process all free vaccines from the schedule, regardless of current child age, and correctly calculate vaccination dates with potential decimal values. A new SQL script is included to create a more accurate database schema for the vaccination management system.

+++ src/main/java/com/vaccination/model/VaccinationScheduleTemplate.java
@@ -6,7 +6,7 @@ public class VaccinationScheduleTemplate {
     private int templateId;
     private int vaccineId;
     private String stageName;
-    private int ageInMonths;
+    private double ageInMonths;
     private int doseNumber;
     private String description;
     private boolean isMandatory;
@@ -18,7 +18,7 @@ public class VaccinationScheduleTemplate {
     public VaccinationScheduleTemplate() {
     }
 
-    public VaccinationScheduleTemplate(int templateId, int vaccineId, String stageName, int ageInMonths, 
+    public VaccinationScheduleTemplate(int templateId, int vaccineId, String stageName, double ageInMonths, 
                                       int doseNumber, String description, boolean isMandatory, int displayOrder) {
         this.templateId = templateId;
         this.vaccineId = vaccineId;
@@ -54,11 +54,11 @@ public class VaccinationScheduleTemplate {
         this.stageName = stageName;
     }
 
-    public int getAgeInMonths() {
+    public double getAgeInMonths() {
         return ageInMonths;
     }
 
-    public void setAgeInMonths(int ageInMonths) {
+    public void setAgeInMonths(double ageInMonths) {
         this.ageInMonths = ageInMonths;
     }
